Remove weird glearray "subproject" artefact
Checking out the glearray repo into this repo didn't work well. Fork marks the folder as a new file with contents "Subproject commit <hash>" but then wont let it be staged. Cloning the repo produces none of the contents of the glearray folder. So killing it, and trying again as a submodule.